% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Segmentation.R
\name{clean_cells}
\alias{clean_cells}
\title{Cleaning function}
\usage{
clean_cells(imthreshold, thr = c(5, 40), verbose = FALSE)
}
\arguments{
\item{imthreshold}{An image of class 'Image'}

\item{thr}{thresholds (lower and upper bound)}

\item{verbose}{Print messages}
}
\description{
Takes an image of class 'Image' as with segmented cells as input,
labels the cells (sets of connected pixels) and applies a threshold
with both a lower and an upper bound to remove cells that are
considered too small or too large (based on area).
}
