% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Segmentation.R
\name{correct_cells}
\alias{correct_cells}
\title{Correction function}
\usage{
correct_cells(im, imfiltered, verbose = FALSE)
}
\arguments{
\item{im}{Image of class 'Image'}

\item{imfiltered}{Image of class 'Image'}

\item{verbose}{Print messages}
}
\description{
Takes an image of class 'Image' and a filtered version of this
image, also of class 'Image' as input. The filtered version is subtracted
from the original image and the intensity values are then normalized to
produce a corrected image.
}
