% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Segmentation.R
\name{filter_cells}
\alias{filter_cells}
\title{Filter function}
\usage{
filter_cells(im, brush.size = 9, verbose = FALSE)
}
\arguments{
\item{im}{image of class 'Image'}

\item{brush.size}{diameter size of brush \link{default: 9 pixels}}

\item{verbose}{Print messages}
}
\description{
Takes an image of class 'Image' as input and applies a
2D FFT convolution filter.
}
