% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Segmentation.R
\name{overlap_fkn}
\alias{overlap_fkn}
\title{Overlap estimate}
\usage{
overlap_fkn(x, y, i)
}
\arguments{
\item{x}{size of x}

\item{y}{size of y}

\item{i}{size of intersect of x and y}
}
\description{
Calculates the overlap of two sets x and y
and their intersect i.
}
