% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Segmentation.R
\name{OverlapImages}
\alias{OverlapImages}
\title{Find overlapping signals}
\usage{
OverlapImages(
  ima,
  imb,
  overlap.min = 0.5,
  return.merged = FALSE,
  return.indices = FALSE,
  verbose = FALSE
)
}
\arguments{
\item{ima}{Image of class 'Image' with segmented data}

\item{imb}{Image of class 'Image' with segmented data}

\item{overlap.min}{Minimum overlap between two cells to consider them the same cell}

\item{return.merged}{Returns overlapping cells but also cells that are strictly non-overlapping}

\item{return.indices}{Also returns the indices of the removed cells}

\item{verbose}{Print messages}
}
\description{
Takes two segmented images "ima" and "imb" of class 'Image' as input and
tries to find cells with overlapping signals by computing shapes with
an overlapping area larger than 'overlap.min'.
}
