% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Segmentation.R
\name{SegmentCells}
\alias{SegmentCells}
\title{Segment cells}
\usage{
SegmentCells(
  impath,
  crop.window = NULL,
  brush.size = 9,
  nsd = 2,
  feature.threshold = c(5, 40),
  tolerance = 0.1,
  return.all = FALSE,
  verbose = FALSE
)
}
\arguments{
\item{impath}{path to immunofluorescence image}

\item{crop.window}{vector of length four specifying a window to crop out
from the provided image. Can be useful if there are artefacts along the edges
of the image.}

\item{brush.size}{diameter size of brush \link{default: 9 pixels}}

\item{nsd}{Number of standard deviations \link{default: 2}}

\item{return.all}{Return images from each step as a list}

\item{verbose}{Print messages}
}
\description{
This function can be used on immunofluorescence images to
segment out stained cells. The segmentation workflow is computed in 5 steps:
\itemize{
\item{filter}
\item{correction}
\item{threshold}
\item{cleaning}
\item{watershed}
}
}
