% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Segmentation.R
\name{threshold_cells}
\alias{threshold_cells}
\title{Threshold function}
\usage{
threshold_cells(imcorrected, nsd = 2, verbose = FALSE)
}
\arguments{
\item{imcorrected}{Image of class 'Image'}

\item{nsd}{Number of standard deviations \link{default: 2}}

\item{verbose}{Print messages}
}
\description{
Takes a "corrected" image of class 'Image' as input and segments the image
by thresholding the intensity values based on 'nsd' number of
standard deviations from the mean intensity.
}
